relayer: moved fetchBlock to a separate module
